in lf buffer, make `yy` yank full path linewise
Previously, `yy` will yank full path to register `"`, without final
newline (e.g. characterwise);
this is because we still needed external program to do basic file
operation, end newline is not expected.

But this causes trouble when use the result inside vim;
and it is not consistent with other linewise operation (like `dd`).

Since we can do cp / mv / rm inside vim (with lf-extend.vim), it is ok
to turn `yy` from characterwise to linewise.

(And, if we need to paste result outside vim, like in a shell, `#` then
paste should suffice.)